client: refactor(auth): move form-login, form-register, and form-forgot-password to src/components/auth and remove unused src/components/layouts
